requirements.txt
aiogram
setuptools
sqlalchemy
psycopg2-binary
pycryptodome